flameclient/tests/utils.py
# -*- coding: utf-8 -*-

# This software is released under the MIT License.
#
# Copyright (c) 2018 Orange Cloud for Business / Cloudwatt
#
# Permission is hereby granted, free of charge, to any person obtaining a copy
# of this software and associated documentation files (the "Software"), to deal
# in the Software without restriction, including without limitation the rights
# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
# copies of the Software, and to permit persons to whom the Software is
# furnished to do so, subject to the following conditions:
#
# The above copyright notice and this permission notice shall be included in
# all copies or substantial portions of the Software.
#
# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
# SOFTWARE.

import six

import openstack

from flameclient.tests import mock
from flameclient.utils import get_deep_attr
from flameclient.utils import munchify


def get_mocked_openstackcloud():
    from flameclient.tests.fixtures import openstackcloud

    cloud = mock.Mock(spec_set=openstack.connection.Connection)
    # for each self.conn.`fixture_module.NAME`() call we will return
    # fixture_module.FIXTURES
    for fixture_module in six.itervalues(openstackcloud.FIXTURES):
        get_deep_attr(
            cloud, fixture_module.NAME
        ).return_value = munchify(fixture_module.FIXTURES)
    return cloud
