requirements.txt
certifi==2018.8.24
futures>=3.1.1,<=3.2.0
keystoneauth1==3.11.0
ndg-httpsclient==0.5.1
netaddr==0.7.19
openstacksdk==0.17.2
os-client-config==1.31.2
pbr==4.2.0
pyasn1==0.4.4
pyOpenSSL==18.0.0
python-cinderclient==4.0.1
python-glanceclient==2.12.1
python-heatclient==1.16.1
python-keystoneclient==3.17.0
python-neutronclient==6.10.0
python-novaclient==11.0.0
python-openstackclient==3.16.1
python-swiftclient==3.6.0
PyYAML==3.13
shade==1.29.0
